modified kafka config
