+++ data-layers/data-layer-kafka/pom.xml
@@ -23,11 +23,7 @@
 			<artifactId>spring-jdbc</artifactId>
 			<version>${org.springframework.version}</version>
 		</dependency>
-        <!-- <dependency>
-          <groupId>org.elasticsearch</groupId>
-          <artifactId>elasticsearch</artifactId>
-          <version>1.3.2</version>
-        </dependency> -->
+
         <dependency>
           <groupId>com.typesafe</groupId>
           <artifactId>config</artifactId>
@@ -45,9 +41,11 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
-			<groupId>org.apache.kafka</groupId>
-			<artifactId>kafka-clients</artifactId>
-			<version>0.8.2-beta</version>
-		</dependency>
-  </dependencies>
+            <groupId>org.apache.kafka</groupId>
+            <artifactId>kafka_2.10</artifactId>
+            <version>0.8.2.1</version>
+        </dependency>
+
+
+ </dependencies>
 </project>
