+++ samples/sample-kafka-client-cluster-consumer/src/main/config.infra.kafka.conf
@@ -2,4 +2,9 @@ zk.connect="localhost:2181"
 serializer.class="kafka.serializer.StringEncoder"
 zk.connectiontimeout.ms=2000
 bootstrap.servers="localhost:9092"
-topic="person"
+payment {
+ 	topic= "payment"
+ 	isSync = true
+ 		}
+
+
